Implemented the ConfigurationInterface and moved the compilerPass to follow the conventions
